[JBMETA-420] Add the EE 9 spec xsds
This also updates XMLResourceResolver to resolve other xsds that are packaged in JBoss Metadata archives but did not have entries in the resolution map. I only did this for ones where the xsd was available under the http://xmlns.jcp.org/xml/ns/javaee/ path. (There were a few that we provide but there was no working URL; those were not added.)

Note that WildFly does not use the JBoss Metadata XMLResourceResolver class.